refactor decomposition.py to factorization.py, update docstrings
