+++ kale/pipeline/multi_domain_adapter.py
@@ -16,9 +16,6 @@ import kale.predict.losses as losses
 from kale.embed.image_cnn import _Bottleneck
 from kale.pipeline.domain_adapter import BaseAdaptTrainer, get_aggregated_metrics
 
-# from torch.optim.lr_scheduler import ReduceLROnPlateau
-# from torch.utils.data import DataLoader, TensorDataset
-
 
 def create_ms_adapt_trainer(method: str, dataset, feature_extractor, task_classifier, **train_params):
     """Methods for multi-source domain adaptation
@@ -345,19 +342,29 @@ class MFSANTrainer(BaseMultiSourceTrainer):
 class _CoIRLS(BaseEstimator, ClassifierMixin):
     """Covariate-Independence Regularized Least Squares (CoIRLS)
 
-    Reference: Zhou, S., 2022. Interpretable Domain-Aware Learning for Neuroimage Classification (Doctoral dissertation,
-        University of Sheffield).
+    Args:
+        kernel (str, optional): {"linear", "rbf", "poly"}. Kernel to use. Defaults to "linear".
+        kernel_kwargs (dict or None, optional): Hyperparameters for the kernel. Defaults to None.
+        alpha (float, optional): Hyperparameter of the l2 (Ridge) penalty. Defaults to 1.0.
+        lambda_ (float, optional): Hyperparameter of the covariate dependence.  Defaults to 1.0.
+
+    Reference:
+        [1] Zhou, S., 2022. Interpretable Domain-Aware Learning for Neuroimage Classification (Doctoral dissertation,
+            University of Sheffield).
+        [2] Zhou, S., Li, W., Cox, C.R., & Lu, H. (2020). Side Information Dependence as a Regularizer for Analyzing
+            Human Brain Conditions across Cognitive Experiments. AAAI 2020, New York, USA.
     """
 
-    def __init__(
-        self, kernel="linear", kernel_kwargs=None, alpha=1.0, lambda_=1.0, max_iter=1000,
-    ):
+    def __init__(self, kernel="linear", kernel_kwargs=None, alpha=1.0, lambda_=1.0):
+        """_summary_
+
+
+        """
         super().__init__()
         self.kernel = kernel
         self.model = None
         self.alpha = alpha
         self.lambda_ = lambda_
-        self.max_iter = max_iter
         if kernel_kwargs is None:
             self.kernel_kwargs = dict()
         else:
@@ -369,6 +376,7 @@ class _CoIRLS(BaseEstimator, ClassifierMixin):
         self.coef_ = None
 
     def fit(self, x, y, covariates):
+
         self._lb.fit(y)
         x = torch.as_tensor(x)
         x = torch.cat([x, torch.ones(x.shape[0], 1)], 1)
@@ -403,6 +411,7 @@ class _CoIRLS(BaseEstimator, ClassifierMixin):
         self.x = x
 
     def predict(self, x):
+
         out = self.decision_function(x)
         if self._lb.y_type_ == "binary":
             pred = self._lb.inverse_transform(torch.sign(out).view(-1))
@@ -412,6 +421,7 @@ class _CoIRLS(BaseEstimator, ClassifierMixin):
         return pred
 
     def decision_function(self, x):
+
         x = torch.as_tensor(x)
         x = torch.cat([x, torch.ones(x.shape[0], 1)], dim=1)
         krnl_x = torch.as_tensor(
